requirements-backend.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.0
pydantic>=2.0.0
python-multipart>=0.0.9
